+++ composeApp/src/commonMain/kotlin/it/homebudget/app/data/csv/import/CsvRowImportHandler.kt
@@ -1,13 +1,11 @@
 package it.homebudget.app.data.csv.import
-
-import com.ionspin.kotlin.bignum.integer.BigInteger
 import it.homebudget.app.data.csv.ParsedUnifiedCsvRow
 
 internal interface CsvRowImportHandler {
     suspend fun importRow(
         row: ParsedUnifiedCsvRow,
         rowIndex: Int,
-        amount: BigInteger,
+        amount: Long,
         itemDate: Long,
         state: CsvImportState
     ): Boolean
